docs: ADRs 0010-0012, three new skill notes, README + CHANGELOG updates
Records the architectural decisions made in this round and indexes the
skill notes worth lifting into other projects.

ADRs:
- 0010 — bearer everywhere with a planted HttpOnly cookie so static UI
  assets fetch without explicit Authorization headers (the project
  requirement that distinguishes june15's stance from fibe-agent's).
- 0011 — UI as an opt-in npm workspace under `ui/`; keeps the runtime
  package light for headless consumers while letting `npm run build`
  produce both halves from a single root install.
- 0012 — image/file attachments delivered as base64 data URLs in the
  message JSON body; sanitized on the way to disk; mentioned to claude
  via `@<absolute-path>` lines.

Skills:
- bearer-on-static-assets.md — the three-source token check + cookie
  carry-over pattern, ready to lift.
- image-attachments-via-data-url.md — the pipeline shape and the
  filename-sanitization that makes it safe.
- e2e-spawn-the-cli.md — separate vitest config, preflight skip,
  spawnGogogo handle, async-generator SSE reader.

README:
- Bearer auth section explains the three accepted carriers and links
  to ADR-0010.
- New "Chat UI (optional)" section walks through enable / serve / dev
  proxy.
- Messages and queue rows show the `attachments` shape; a curl example
  illustrates inline base64 from `photo.png`.
- Contributing note covers `npm run test:e2e` and its skip behavior.

CHANGELOG:
- Unreleased section enumerates the new features.

Some lint fixups from the surrounding files came along: dot-notation
for env reads in the e2e helpers, a missing UploadStore type-import in
the routes test, sse-client's reader.read() typed as Uint8Array, and a
narrow eslint-disable around an intentional unconditional truthy check
in the e2e afterAll.

+++ tests/e2e/helpers/sse-client.ts
@@ -13,7 +13,7 @@ export async function* readSse(res: Response): AsyncGenerator<SseFrame> {
   for (;;) {
     const { value, done } = await reader.read();
     if (done) return;
-    buf += decoder.decode(value, { stream: true });
+    buf += decoder.decode(value as Uint8Array, { stream: true });
     let sep: number;
     while ((sep = buf.indexOf('\n\n')) !== -1) {
       const raw = buf.slice(0, sep);
